rename remote and local packages to data_source, also add details and favourites fragment
