+++ .idea/deploymentTargetDropDown.xml
@@ -12,6 +12,6 @@
         </deviceKey>
       </Target>
     </targetSelectedWithDropDown>
-    <timeTargetWasSelectedWithDropDown value="2022-02-17T07:52:49.642178Z" />
+    <timeTargetWasSelectedWithDropDown value="2022-02-19T10:47:48.250631Z" />
   </component>
 </project>
