+++ app/src/main/java/ge/example/githubapidemo/feature_github_repositories/presentation/fragment/details/DetailsFragment.kt
@@ -1,0 +1,12 @@
+package ge.example.githubapidemo.feature_github_repositories.presentation.fragment.details
+
+import ge.example.githubapidemo.databinding.FragmentDetailsBinding
+import ge.example.githubapidemo.feature_github_repositories.presentation.fragment.BaseFragment
+
+
+class DetailsFragment : BaseFragment<FragmentDetailsBinding>(FragmentDetailsBinding::inflate) {
+    override fun start() {
+    }
+
+
+}
