+++ app/src/main/java/ge/example/githubapidemo/feature_github_repositories/presentation/fragment/favourite/FavouritesFragment.kt
@@ -1,0 +1,13 @@
+package ge.example.githubapidemo.feature_github_repositories.presentation.fragment.favourite
+
+import ge.example.githubapidemo.databinding.FragmentFavouritesBinding
+import ge.example.githubapidemo.feature_github_repositories.presentation.fragment.BaseFragment
+
+
+class FavouritesFragment :
+    BaseFragment<FragmentFavouritesBinding>(FragmentFavouritesBinding::inflate) {
+    override fun start() {
+    }
+
+
+}
